String/File refactoring in src/qt/
