+++ contrib/oasis-qt.pro
@@ -168,7 +168,7 @@ HEADERS += src/activemasternode.h \
            src/qt/addresstablemodel.h \
            src/qt/askpassphrasedialog.h \
            src/qt/bitcoinaddressvalidator.h \
-           src/qt/bitcoinamountfield.h \
+           src/qt/oasisamountfield.h \
            src/qt/bitcoingui.h \
            src/qt/bitcoinunits.h \
            src/qt/clientmodel.h \
@@ -332,7 +332,7 @@ HEADERS += src/activemasternode.h \
            src/leveldb/port/win/stdint.h \
            src/secp256k1/src/java/org_bitcoin_NativeSecp256k1.h \
            src/crypto/aes_helper.c \
-           src/qt/bitcoinamountfield.moc \
+           src/qt/oasisamountfield.moc \
            src/qt/oasis.moc \
            src/qt/intro.moc \
            src/qt/overviewpage.moc \
@@ -463,9 +463,9 @@ SOURCES += src/activemasternode.cpp \
            src/qt/addresstablemodel.cpp \
            src/qt/askpassphrasedialog.cpp \
            src/qt/bitcoinaddressvalidator.cpp \
-           src/qt/bitcoinamountfield.cpp \
-           src/qt/bitcoingui.cpp \
-           src/qt/bitcoinunits.cpp \
+           src/qt/oasisamountfield.cpp \
+           src/qt/oasisgui.cpp \
+           src/qt/oasisunits.cpp \
            src/qt/clientmodel.cpp \
            src/qt/coincontroldialog.cpp \
            src/qt/coincontroltreewidget.cpp \
