+++ src/qt/forms/receivecoinsdialog.ui
@@ -353,7 +353,7 @@
   <customwidget>
    <class>BitcoinAmountField</class>
    <extends>QLineEdit</extends>
-   <header>bitcoinamountfield.h</header>
+   <header>oasisamountfield.h</header>
    <container>1</container>
   </customwidget>
  </customwidgets>
