+++ src/qt/forms/sendcoinsdialog.ui
@@ -1451,7 +1451,7 @@
   <customwidget>
    <class>BitcoinAmountField</class>
    <extends>QLineEdit</extends>
-   <header>bitcoinamountfield.h</header>
+   <header>oasisamountfield.h</header>
    <container>1</container>
   </customwidget>
  </customwidgets>
