+++ src/qt/locale/oasis_en.ts
@@ -505,7 +505,7 @@
 <context>
     <name>BitcoinGUI</name>
     <message>
-        <location filename="../bitcoingui.cpp" line="+124"/>
+        <location filename="../oasisgui.cpp" line="+124"/>
         <source>Wallet</source>
         <translation>Wallet</translation>
     </message>
@@ -3356,7 +3356,7 @@ Either mint higher denominations (so fewer inputs are needed) or reduce the amou
 <context>
     <name>QObject</name>
     <message>
-        <location filename="../bitcoinunits.cpp" line="+252"/>
+        <location filename="../oasisunits.cpp" line="+252"/>
         <source>Amount</source>
         <translation>Amount</translation>
     </message>
@@ -5586,7 +5586,7 @@ Either mint higher denominations (so fewer inputs are needed) or reduce the amou
 <context>
     <name>UnitDisplayStatusBarControl</name>
     <message>
-        <location filename="../bitcoingui.cpp" line="+102"/>
+        <location filename="../oasisgui.cpp" line="+102"/>
         <source>Unit to show amounts in. Click to select another unit.</source>
         <translation>Unit to show amounts in. Click to select another unit.</translation>
     </message>
