+++ src/qt/oasisamountfield.cpp
@@ -2,9 +2,9 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "bitcoinamountfield.h"
+#include "oasisamountfield.h"
 
-#include "bitcoinunits.h"
+#include "oasisunits.h"
 #include "guiconstants.h"
 #include "qvaluecombobox.h"
 
@@ -177,7 +177,7 @@ signals:
     void valueChanged();
 };
 
-#include "bitcoinamountfield.moc"
+#include "oasisamountfield.moc"
 
 BitcoinAmountField::BitcoinAmountField(QWidget* parent) : QWidget(parent),
                                                           amount(0)
