+++ src/qt/oasisunits.cpp
@@ -4,7 +4,7 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "bitcoinunits.h"
+#include "oasisunits.h"
 #include "chainparams.h"
 #include "primitives/transaction.h"
