+++ src/qt/overviewpage.cpp
@@ -8,7 +8,7 @@
 #include "overviewpage.h"
 #include "ui_overviewpage.h"
 
-#include "bitcoinunits.h"
+#include "oasisunits.h"
 #include "clientmodel.h"
 #include "guiconstants.h"
 #include "guiutil.h"
