+++ src/qt/transactiontablemodel.h
@@ -5,7 +5,7 @@
 #ifndef BITCOIN_QT_TRANSACTIONTABLEMODEL_H
 #define BITCOIN_QT_TRANSACTIONTABLEMODEL_H
 
-#include "bitcoinunits.h"
+#include "oasisunits.h"
 
 #include <QAbstractTableModel>
 #include <QStringList>
